+++ package/sdk/RELEASES.md
@@ -288,3 +288,13 @@ Based on:
 - [typescript v0.4.0] package/sdk
 ### Releases
 - [NPM v0.4.0] https://www.npmjs.com/package/ryan-total-test-act/v/0.4.0 - package/sdk
+
+## 2025-02-20 17:07:10
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.498.0 (2.518.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.6.0] package/sdk
+### Releases
+- [NPM v0.6.0] https://www.npmjs.com/package/ryan-total-test-act/v/0.6.0 - package/sdk